requirements.txt
dlt[duckdb]>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0
pytest>=7.0.0
pandas>=1.3.0
python-dateutil>=2.8.2
pytest-mock>=3.10.0
pytest-cov>=3.0.0
